refactor(ui): split SettingsScreen god-file into per-domain sections
SettingsScreen.kt was 1062 LOC: orchestrator + form state +
helpers + Interface/Behavior/Network/Experimental/DataDir/
Diagnostics/About blocks all inline. Each domain block was
~150 LOC of imperative item { ... } code that could only be
read or modified by scrolling through the whole file.

Split into a thin orchestrator + 5 section composables + 2
support files. Each new file has a single domain responsibility
and is small enough to read in one sitting:

  SettingsScreen.kt         137  orchestrator: form, save(),
                                 LazyColumn calling sections
  AppearanceSection.kt      269  language, theme picker,
                                 background, dark theme,
                                 HomeView, UiStyle, close-after,
                                 offline mode
  NetworkSection.kt         141  SSL bypass list (live) + revoke
                                 + force-proxy toggle
  ExperimentalSection.kt    216  master + mandatory updates +
                                 prerelease + autosync + jvm
                                 builder + mirror + mimic version
                                 (with debounced text input)
  AdvancedSection.kt        170  data-dir move (picker + confirm
                                 dialog + DataDirMover.schedule
                                 + exitProcess on commit)
  DiagnosticsSection.kt     237  logs/crash folders, diag bundle
                                 generator, GitHub issue prefill,
                                 AprilFools debug-panel reveal,
                                 About card
  SettingsFormState.kt       70  per-toggle mutableState +
                                 mergeInto onto SettingsData;
                                 internal so sibling sections
                                 read/write directly
  SettingsHelpers.kt        216  shared composables: section
                                 title, switch-row with icon,
                                 plain switch-row, HomeViewPicker
                                 + UiStylePicker pill widgets

Behavior unchanged: same save() flow (mirror to NetworkState +
Protocol.setMimicLauncherVersion), same persistence shape, same
puppet selectors. Just structurally legible.

Atelier-side visual rework now operates on small files instead
of a single monolith.
